Xamarin Android App for displaying carts and cart details
A simple Android app developed using Xamarin where list of cart items
are being showed and details of each cart. Here the main thing is Custom
Adapter is being used to display cart details which has a view which
needs to be displayed below listview which can which be achieved by
Custom Adapter.